Confirmation alert for delete user

+++ frontend/src/routes/admin/+page.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import { isAuthenticated, user } from '../../auth/auth';
 	import ErrorAdmin from '../../components/ErrorAdmin.svelte';
+	import Confirmation from '../../components/Confirmation.svelte';
 	export let data;
 	const userName = data.props.userName;
 	const userId = data.props.userId;
@@ -8,20 +9,41 @@
 	const dataUsers = data.props.dataUsers;
 	user.set(userName);
 
-	async function deleteUser(id: number) {
-		try {
-			const url = `http://localhost:8080/auth/admin/${id}`;
-			const response = await fetch(url, {
-				method: 'DELETE',
-				credentials: 'include'
-			});
+	let showConfirmation = false;
+	let userIdToDelete: number | null = null;
+	let usernameToDelete: string | null = null;
 
-			if (response.ok) {
-				location.reload();
+	function askDeleteConfirmation(id: number) {
+		userIdToDelete = id;
+		usernameToDelete = dataUsers.find(
+			(user: { id: number }) => user.id === userIdToDelete
+		)?.username;
+		showConfirmation = true;
+	}
+
+	async function deleteUser() {
+		if (userIdToDelete !== null) {
+			try {
+				const url = `http://localhost:8080/auth/admin/${userIdToDelete}`;
+				const response = await fetch(url, {
+					method: 'DELETE',
+					credentials: 'include'
+				});
+
+				if (response.ok) {
+					location.reload();
+				}
+			} catch (error) {
+				console.error('Fetch error:', error);
 			}
-		} catch (error) {
-			console.error('Fetch error:', error);
 		}
+		showConfirmation = false; // Close the confirmation dialog
+		userIdToDelete = null; // Reset the deletion ID
+	}
+
+	function cancelDelete() {
+		showConfirmation = false;
+		userIdToDelete = null;
 	}
 </script>
 
@@ -34,51 +56,37 @@
 	<div class="container mx-auto py-8">
 		<h1 class="text-3xl text-center mb-4">Admin page</h1>
 		<p class="text-lg text-gray-700 leading-relaxed">Welcome, {userName}!</p>
-		<p class="text-lg text-gray-700 leading-relaxed">
-			Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed ut ultricies nisl. Phasellus
-			ultricies sem a massa ultricies, ac tincidunt est laoreet. Proin pulvinar felis at sem
-			malesuada, et ultricies purus posuere.
-		</p>
 		<table class="min-w-full leading-normal mt-4">
 			<thead>
 				<tr>
 					<th
 						class="px-5 py-3 border-b-2 border-gray-200 bg-gray-100 text-left text-xs font-semibold text-gray-600 uppercase tracking-wider"
+						>Username</th
 					>
-						Username
-					</th>
 					<th
 						class="px-5 py-3 border-b-2 border-gray-200 bg-gray-100 text-left text-xs font-semibold text-gray-600 uppercase tracking-wider"
+						>Email</th
 					>
-						Email
-					</th>
 					<th
 						class="px-5 py-3 border-b-2 border-gray-200 bg-gray-100 text-left text-xs font-semibold text-gray-600 uppercase tracking-wider"
+						>Role</th
 					>
-						Role
-					</th>
 					<th
 						class="px-5 py-3 border-b-2 border-gray-200 bg-gray-100 text-left text-xs font-semibold text-gray-600 uppercase tracking-wider"
+						>Actions</th
 					>
-						Actions
-					</th>
 				</tr>
 			</thead>
 			<tbody>
 				{#each dataUsers as user}
 					<tr>
+						<td class="px-5 py-2 border-b border-gray-200 bg-white text-sm">{user.username}</td>
+						<td class="px-5 py-2 border-b border-gray-200 bg-white text-sm">{user.email}</td>
+						<td class="px-5 py-2 border-b border-gray-200 bg-white text-sm">{user.role}</td>
 						<td class="px-5 py-2 border-b border-gray-200 bg-white text-sm">
-							{user.username}
-						</td>
-						<td class="px-5 py-2 border-b border-gray-200 bg-white text-sm">
-							{user.email}
-						</td>
-						<td class="px-5 py-2 border-b border-gray-200 bg-white text-sm">
-							{user.role}
-						</td>
-						<td class="px-5 py-2 border-b border-gray-200 bg-white text-sm">
-							<button on:click={() => deleteUser(user.id)} class="text-red-500 hover:text-red-700"
-								>Delete</button
+							<button
+								on:click={() => askDeleteConfirmation(user.id)}
+								class="text-red-500 hover:text-red-700">Delete</button
 							>
 						</td>
 					</tr>
@@ -86,6 +94,12 @@
 			</tbody>
 		</table>
 	</div>
+	<Confirmation
+		isOpen={showConfirmation}
+		message="Are you sure you want to delete {usernameToDelete}?"
+		onConfirm={deleteUser}
+		onCancel={cancelDelete}
+	/>
 {:else}
 	<ErrorAdmin />
 {/if}
